Change mismatch of bytes 11 to 90 to be a warning, not an error, and print out the values of the bytes that do not match. Add comment explaining that there is no documented rationale for the check.
fsck_msdos checks that several bytes are equal in the primary and
backup bootblock.  There is no documented rationale.  The kernel does
not enforce this when mounting a filesystem.  I have a FAT32
filesystem on my disk (mounted as /share) which fails this check, but
is accepted by both Windows XP and NetBSD.  My volume differs in bytes
71-81.  The primary contains "SHARE " and the backup NUL bytes.

Further, fsck_msdos does not have code to repair the mismatch.  As a
result such filesystems can be used not checked.