Add a NOTES section to the manual, indicating that kauth(9) is still under active development and its ABI (and possibly API) may change between NetBSD versions.
This is critical to, for example, LKMs, where there might be a case of them
being built using one version of the ABI and used on system with another.

The main concern for "ABI" here is the set of KAUTH_* actions and requests
that is (for now) an enum. This note is likely to be removed as kauth(9)
is stablized -- hopefully before NetBSD 5.0.

okay christos@