From: Andy Shevchenko <andy.shevchenko@teleca.com>
The OMAP2 ref timer providing a base for generic timecounting was configured
to reload at HZ similarly to the clock providing the tick interrupt at HZ.
This and the fact that the tc framework expected a full 32 bit counter range
resulted in time running ~40 times faster than normal.

Configure the ref timer as a full range counter to fix this.